Add actual signature validation
The signature is checked using the System.Security.Cryptography.Pkcs
package. This package unfortunately doesn't support .NET Framework 4.0
which is why I had to drop support for that target framework.

A test case ensures that the verification actually works.
It uses a manipulated binary where the signature shouldn't be
valid any more.
I simply changed 1 bit in the binary and then exchanged the included
hash so it matches the changed binary. The previous verification
reported the signature as valid as it only compared the hashes.

Unfortunately, two other test cases are failing now where the signature
is reported as invalid, but it should be valid.

Fixes #59